+++ src/main/java/sylenthuntress/aceofhearts/event/GracePeriod.java
@@ -1,5 +1,6 @@
 package sylenthuntress.aceofhearts.event;
 
+import net.fabricmc.fabric.api.entity.event.v1.ServerEntityWorldChangeEvents;
 import net.fabricmc.fabric.api.entity.event.v1.ServerLivingEntityEvents;
 import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
 import net.minecraft.entity.Entity;
@@ -7,14 +8,16 @@ import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.registry.RegistryKeys;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
+import net.minecraft.world.dimension.DimensionTypes;
 import sylenthuntress.aceofhearts.registry.ModAttachmentTypes;
 
 @SuppressWarnings("UnstableApiUsage")
-public class GracePeriod implements ServerLivingEntityEvents.AllowDamage, ServerPlayerEvents.Join, ServerPlayerEvents.AfterRespawn {
+public class GracePeriod implements ServerEntityWorldChangeEvents.AfterPlayerChange, ServerLivingEntityEvents.AllowDamage, ServerPlayerEvents.Join, ServerPlayerEvents.AfterRespawn {
     @Override
     public void afterRespawn(ServerPlayerEntity oldPlayer, ServerPlayerEntity newPlayer, boolean alive) {
         newPlayer.setAttached(ModAttachmentTypes.GRACE_PERIOD, 6000);
@@ -52,5 +55,13 @@ public class GracePeriod implements ServerLivingEntityEvents.AllowDamage, Server
         }
 
         return true;
+    }
+
+    @Override
+    public void afterChangeWorld(ServerPlayerEntity player, ServerWorld oldWorld, ServerWorld newWorld) {
+        var registry = newWorld.getRegistryManager().getOrThrow(RegistryKeys.DIMENSION_TYPE);
+        if (registry.get(DimensionTypes.THE_END) == newWorld.getDimension()) {
+            player.modifyAttached(ModAttachmentTypes.GRACE_PERIOD, duration -> duration + 600);
+        }
     }
 }